refactor: resolve theme merge and modularize helpers
